Int as arg type for ID

+++ frontend/src/graphql/mutations.js
@@ -73,13 +73,13 @@ export const createTransactionLink = gql`
 `
 
 export const deleteTransactionLink = gql`
-  mutation($id: Float!) {
+  mutation($id: Int!) {
     deleteTransactionLink(id: $id)
   }
 `
 
 export const redeemTransactionLink = gql`
-  mutation($id: Float!) {
+  mutation($id: Int!) {
     redeemTransactionLink(id: $id)
   }
 `
